+++ crates/nextchat-server/src/lib.rs
@@ -1,0 +1,13 @@
+//! NextChat Server library.
+
+use nextchat_database::Client;
+
+mod controllers;
+mod response;
+mod services;
+
+pub async fn run(client: &Client, host: [u8; 4], port: u16) {
+    warp::serve(controllers::routes(client))
+        .run((host, port))
+        .await;
+}