+++ crates/nextchat-utils/src/lib.rs
@@ -1,0 +1,6 @@
+//! NextChat Utils library.
+
+mod versions;
+pub use versions::{AppVersion, AppVersions};
+
+pub use semver::Version;