+++ src/models.rs
@@ -1,5 +1,0 @@
-//! NextChat Server models module.
-
-pub mod connection;
-pub mod friends;
-pub mod users;