+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased (0.1.0-alpha1)
 
+### 22/03/2021
+-   Add NextChat Communication module.
+-   Add a new modular structure.
+
 ### 20/03/2021
 -   Code refactor to a scalable structure.
 -   Add code documentation.